Move FAB to Plaza so it can be animated
