--- mikado-app/src/components/Overlays/Overlay.js
@@ -1,12 +1,0 @@
-import FAB from "./FAB.js"
-
-function Overlay({ FABonClick }) {
-  return (
-
-    <FAB onClick={FABonClick} />
-
-  );
-
-}
-
-export default Overlay;
